validate keyring controller migration
